pyproject.toml
[project]
name = "green-agent-template"
version = "0.1.0"
description = "A template for A2A green agents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "agentbeats>=1.2.6",
    "pydantic>=2.4.0",
    "uvicorn>=0.38.0",
    "fastapi>=0.100.0",
    "requests>=2.31.0",
    "aiohttp>=3.13.3",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
]
